refactor: extract pages into multiple files
