+++ websites/M/MyFigureCollection.net/pages/PicturePage.ts
@@ -1,0 +1,35 @@
+import { BACKGROUND_URL_REGEX, getButton, getCurrentLink, slideshow } from '../util.js'
+import { BasePage, strings } from './index.js'
+
+export class PicturePage extends BasePage {
+  override async executeBrowse(presenceData: PresenceData): Promise<boolean> {
+    presenceData.details = strings.viewList
+    presenceData.state = getCurrentLink()
+    const pictures = document.querySelectorAll('.results .picture-icon')
+    for (const picture of pictures) {
+      const data = { ...presenceData }
+      const link = picture.querySelector('a')?.href ?? document.location.href
+      data.smallImageKey = picture.querySelector('span')?.style.background.match(BACKGROUND_URL_REGEX)?.[1]
+      data.buttons = [{ label: strings.buttonViewPicture, url: link }]
+      slideshow.addSlide(link, data, MIN_SLIDE_TIME)
+    }
+    return true
+  }
+
+  override async executeView(presenceData: PresenceData, id: string): Promise<boolean> {
+    presenceData.details = strings.viewPicture
+    presenceData.state = `#${id}`
+    presenceData.smallImageKey = document.querySelector<HTMLImageElement>('.the-picture img')
+    presenceData.smallImageText = strings.byAuthor.replace(
+      '{author}',
+      document.querySelector('.object-meta .user-anchor')?.textContent ?? 'Unknown',
+    )
+    presenceData.buttons = [getButton(strings.buttonViewPicture)]
+    const relatedItem = document.querySelector('.tbx-target-ITEMS .stamp')
+    if (relatedItem) {
+      presenceData.state += ` - ${relatedItem.textContent}`
+      presenceData.buttons.push({ label: strings.viewItem, url: relatedItem.querySelector('a') })
+    }
+    return false
+  }
+}
