+++ websites/M/MyFigureCollection.net/pages/TagPage.ts
@@ -1,0 +1,10 @@
+import { getTitle } from '../util.js'
+import { BasePage, strings } from './index.js'
+
+export class TagPage extends BasePage {
+  override async executeView(presenceData: PresenceData, _id: string): Promise<boolean> {
+    presenceData.details = strings.browseTag
+    presenceData.state = getTitle()
+    return false
+  }
+}
